Server/requirements.txt
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
pytz>=2023.3
geopy>=2.3.0
python-dotenv>=1.0.0
